+++ src/components/headings/GeneralHeader.js
@@ -12,6 +12,7 @@ import Navbar from 'components/common/Navbar';
 
 import { authRoles } from 'auth';
 import appLogo from 'assets/images/logo.png';
+import appLogoWhite from 'assets/images/logo-white.png';
 
 const useStyles = makeStyles((theme) => ({
     fixed: {
@@ -50,14 +51,21 @@ const useStyles = makeStyles((theme) => ({
 
 export default function GeneralHeader({ fixed }) {
     const classes = useStyles();
+
     const menuItems = useSelector(({ ui }) => ui.sidebar.menuItems);
     const appInformation = useSelector(({ ui }) => ui.appInformation);
     const { authenticated, data } = useSelector(({ auth }) => auth.user);
+    const darkTheme = useSelector(({ ui }) => ui.appSettings?.theme) === 'dark';
 
     return (
         <Grid component="header" container className={clsx(classes.root, fixed && classes.fixed)}>
             <Grid container className={classes.logoWrapper} item xs={6}>
-                <Logo className={classes.logo} size="small" title={appInformation?.appTitle} imageSrc={appLogo} />
+                <Logo
+                    className={classes.logo}
+                    size="small"
+                    title={appInformation?.appTitle}
+                    imageSrc={darkTheme ? appLogoWhite : appLogo}
+                />
                 {authenticated && data?.role === authRoles.patient && (
                     <CallModule classes={{ button: classes.callButton }} />
                 )}