+++ src/config/languages/spanishContent.js
@@ -169,6 +169,15 @@ const spanishContent = {
             }
         }
     },
+    projectPage: {
+        overviewLabel: 'Overview',
+        focusLabel: 'Focus',
+        categoriesLabel: 'Categories',
+        briefLabel: 'Brief',
+        researchLabel: 'Research',
+        visualConceptLabel: 'Visual Concept',
+        prototypeLabel: 'Prototype'
+    },
     settingsPage: {
         menu: {
             title: 'Configuración',