+++ src/data/shakaApp.js
@@ -1,0 +1,5 @@
+const shakaApp = {
+    title: 'Shaka App'
+};
+
+export default shakaApp;