+++ src/pages/home/Banner.js
@@ -19,7 +19,7 @@ const useStyles = makeStyles((theme) => ({
         width: '50%'
     },
     title: {
-        color: theme.palette.primary.main
+        color: theme.palette.type === 'dark' ? theme.palette.primary.main : theme.palette.text.primary
     }
 }));