+++ changelog.md
@@ -3,3 +3,4 @@
 -   Feature: Start Project from TS Template Project.
 -   Feature: Add basic structure and Banner.
 -   Feature: Add about and work sections.
+-   Feature: Add Project Page and projects slice.